requirements_2.txt
Flask
mysql_connector
paho_mqtt
